Split utils into to separate files
