+++ src/ts/arrayutils.ts
@@ -1,0 +1,17 @@
+export namespace ArrayUtils {
+  /**
+   * Removes an item from an array.
+   * @param array the array that may contain the item to remove
+   * @param item the item to remove from the array
+   * @returns {any} the removed item or null if it wasn't part of the array
+   */
+  export function remove<T>(array: T[], item: T): T | null {
+    let index = array.indexOf(item);
+
+    if (index > -1) {
+      return array.splice(index, 1)[0];
+    } else {
+      return null;
+    }
+  }
+}
