+++ src/ts/components/controlbar.ts
@@ -1,6 +1,6 @@
 import {ContainerConfig, Container} from './container';
 import {UIInstanceManager} from '../uimanager';
-import {UIUtils} from '../utils';
+import {UIUtils} from '../uiutils';
 import {Spacer} from './spacer';
 
 /**
