+++ src/ts/components/recommendationoverlay.ts
@@ -2,7 +2,7 @@ import {ContainerConfig, Container} from './container';
 import {Component, ComponentConfig} from './component';
 import {DOM} from '../dom';
 import {UIInstanceManager, UIRecommendationConfig} from '../uimanager';
-import {StringUtils} from '../utils';
+import {StringUtils} from '../stringutils';
 import {HugeReplayButton} from './hugereplaybutton';
 
 /**
