+++ src/ts/components/subtitlesettings/fontfamilyselectbox.ts
@@ -1,6 +1,6 @@
 import {SubtitleSettingSelectBoxConfig, SubtitleSettingSelectBox} from './subtitlesettingselectbox';
 import {UIInstanceManager} from '../../uimanager';
-import {StorageUtils} from '../../utils';
+import {StorageUtils} from '../../storageutils';
 
 interface FontFamily {
   name: string;
