+++ src/ts/components/subtitlesettings/fontsizeselectbox.ts
@@ -1,6 +1,6 @@
 import {SubtitleSettingSelectBoxConfig, SubtitleSettingSelectBox} from './subtitlesettingselectbox';
 import {UIInstanceManager} from '../../uimanager';
-import {StorageUtils} from '../../utils';
+import {StorageUtils} from '../../storageutils';
 
 /**
  * A select box providing a selection of different font colors.
